# src/main/java/assignment/bank/rest/AccountCreateRequest.java
package assignment.bank.rest;

public class AccountCreateRequest {
	
	private double initialDeposit;
	
	public double getInitialDeposit() {
		return initialDeposit;
	}

	public void setInitialDeposit(double initialDeposit) {
		this.initialDeposit = initialDeposit;
	}
}