# src/main/java/assignment/bank/rest/AccountDepositRequest.java
package assignment.bank.rest;

public class AccountDepositRequest {

	private double amount;

	public double getAmount() {
		return amount;
	}
	public void setAmount(double amount) {
		this.amount = amount;
	}
}