# src/main/java/assignment/bank/rest/AccountTransferRequest.java
package assignment.bank.rest;

public class AccountTransferRequest {
	
	private String toAccountNumber;
	
	private Double amount;

	public String getToAccountNumber() {
		return toAccountNumber;
	}

	public void setToAccountNumber(String toAccountNumber) {
		this.toAccountNumber = toAccountNumber;
	}

	public Double getAmount() {
		return amount;
	}

	public void setAmount(double amount) {
		this.amount = amount;
	}
}