# src/main/java/assignment/bank/rest/ErrorResponse.java
package assignment.bank.rest;

public class ErrorResponse {
	
	private String message;
	
	public ErrorResponse(String message) {
		this.message = message;
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}
}